etl/requirements.txt
pandas==2.2.2
pyyaml
psycopg2-binary
clickhouse-driver
python-dateutil
sqlalchemy
